+++ src/main/java/io/github/astrapi69/swing/action/ExitApplicationAction.java
@@ -22,53 +22,30 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package io.github.astrapi69.swing.actions;
+package io.github.astrapi69.swing.action;
 
 import java.awt.event.ActionEvent;
 
 import javax.swing.*;
 
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NonNull;
-import lombok.experimental.FieldDefaults;
-
 /**
- * The abstract class {@link ShowFrameAction}
+ * The class {@link ExitApplicationAction} finish the application
  */
-@Getter
-@FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
-public abstract class ShowFrameAction extends AbstractAction
+public class ExitApplicationAction extends AbstractAction
 {
 
+	/** The Constant serialVersionUID. */
 	private static final long serialVersionUID = 1L;
 
-	/** The title of the frame */
-	String title;
-
 	/**
-	 * Instantiates a new {@link ShowFrameAction} object
+	 * Instantiates a new {@link ExitApplicationAction} object
 	 *
 	 * @param name
 	 *            the name
 	 */
-	public ShowFrameAction(final @NonNull String name)
-	{
-		this(name, name);
-	}
-
-	/**
-	 * Instantiates a new {@link ShowFrameAction} object
-	 *
-	 * @param name
-	 *            the name
-	 * @param title
-	 *            the title
-	 */
-	public ShowFrameAction(final String name, final @NonNull String title)
+	public ExitApplicationAction(final String name)
 	{
 		super(name);
-		this.title = title;
 	}
 
 	/**
@@ -77,15 +54,17 @@ public abstract class ShowFrameAction extends AbstractAction
 	@Override
 	public void actionPerformed(final ActionEvent e)
 	{
-		onShowFrame(e);
+		onExit(e);
 	}
 
 	/**
-	 * Abstract callback method to interact when the Frame will show
+	 * Callback method to interact on exit
 	 *
-	 * @param e
+	 * @param actionEvent
 	 *            the action event
 	 */
-	protected abstract void onShowFrame(final ActionEvent e);
-
+	protected void onExit(final ActionEvent actionEvent)
+	{
+		System.exit(0);
+	}
 }