+++ CHANGELOG.md
@@ -20,6 +20,7 @@ CHANGED:
 - update of com.github.ben-manes.versions.gradle.plugin to new version 0.41.0
 - update of jobj-core dependency to new version 5.2
 - update of model-data dependency to new version 1.12
+- renamed package actions to action
 
 Version 1.3
 -------------